refactor(archive): move raw payloads out of lib
