+++ docs/topology-status.md
@@ -11,12 +11,14 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 | Owner | Title | Cells | At target | Pending | Conflicts |
 | --- | --- | --: | --: | --: | --: |
+| archive-action-event | — | 4 | 4 | 0 | 0 |
 | archive-filter | archive filter semantics | 4 | 4 | 0 | 0 |
 | archive-query | archive query semantics | 17 | 17 | 0 | 0 |
+| archive-raw-payload | — | 5 | 5 | 0 | 0 |
 
 ### Summary
 
-- **Stable** (no move scoped): 637
+- **Stable** (no move scoped): 628
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (lib-root + storage-root): 29
 - **TBD** (cell needs explicit assignment): 0
