+++ polylogue/archive/raw_payload/__init__.py
@@ -2,20 +2,20 @@
 
 from __future__ import annotations
 
-from polylogue.lib.json import JSONDocument, JSONDocumentList, JSONScalar, JSONValue
-from polylogue.lib.raw_payload.decode import (
+from polylogue.archive.raw_payload.decode import (
     RawPayloadEnvelope,
     WireFormat,
     build_raw_payload_envelope,
     sample_jsonl_payload,
 )
-from polylogue.lib.raw_payload.sampling_buckets import is_record_candidate, record_bucket_key
-from polylogue.lib.raw_payload.sampling_extract import (
+from polylogue.archive.raw_payload.sampling_buckets import is_record_candidate, record_bucket_key
+from polylogue.archive.raw_payload.sampling_extract import (
     collect_limited_samples,
     extract_payload_samples,
     extract_record_samples_from_raw_content,
     limit_samples,
 )
+from polylogue.lib.json import JSONDocument, JSONDocumentList, JSONScalar, JSONValue
 
 __all__ = [
     "JSONDocument",
