+++ polylogue/schemas/sampling_db.py
@@ -8,14 +8,14 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Literal, TypeAlias, overload
 
+from polylogue.archive.raw_payload import extract_record_samples_from_raw_content
+from polylogue.archive.raw_payload.decode import RawPayloadEnvelope
 from polylogue.lib.json import JSONDocument
 from polylogue.lib.provider_identity import (
     CORE_RUNTIME_PROVIDERS,
     canonical_runtime_provider,
     canonical_schema_provider,
 )
-from polylogue.lib.raw_payload import extract_record_samples_from_raw_content
-from polylogue.lib.raw_payload.decode import RawPayloadEnvelope
 from polylogue.logging import get_logger
 from polylogue.paths import db_path as archive_db_path
 from polylogue.schemas.observation import (
