+++ polylogue/schemas/synthetic/core.py
@@ -25,7 +25,7 @@ from polylogue.schemas.synthetic.semantic_values import SemanticValueGenerator
 from polylogue.schemas.synthetic.wire_formats import WireFormat
 
 if TYPE_CHECKING:
-    from polylogue.lib.raw_payload.decode import JSONValue
+    from polylogue.archive.raw_payload.decode import JSONValue
     from polylogue.schemas.synthetic.showcase import ConversationTheme
