+++ polylogue/schemas/synthetic/runtime.py
@@ -8,7 +8,7 @@ import uuid
 from datetime import datetime, timezone
 from typing import TYPE_CHECKING, Protocol, TypeAlias
 
-from polylogue.lib.raw_payload.decode import JSONValue
+from polylogue.archive.raw_payload.decode import JSONValue
 from polylogue.schemas.synthetic.models import SchemaRecord, SchemaValue
 from polylogue.schemas.synthetic.semantic_values import SemanticValueGenerator, _text_for_role
 from polylogue.schemas.synthetic.wire_formats import WireFormat
