+++ tests/unit/pipeline/test_parsing_service.py
@@ -12,9 +12,9 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
+from polylogue.archive.raw_payload.decode import JSONValue
 from polylogue.config import Config, Source
 from polylogue.errors import DatabaseError
-from polylogue.lib.raw_payload.decode import JSONValue
 from polylogue.pipeline.payload_types import ParseBatchObservation
 from polylogue.pipeline.services.acquisition import AcquireResult, AcquisitionService
 from polylogue.pipeline.services.acquisition_records import ScanResult
